Add files via upload
1mg Homeopathic Data Analysis and Dashboard Creation

The project aims to utilize the Python library Beautifulsoup to scrape data from the renowned online medicine delivery platform, 1mg. The specific objective is to gather comprehensive information on homeopathic medicines available on the platform. This data will then be analyzed to generate valuable insights and recommendations for a consultancy firm's client who intends to establish a homeopathic medicine store.

By extracting data such as medicine names, prices, descriptions, and customer ratings, the project will enable a comprehensive understanding of the market landscape. This information will be instrumental in estimating the total cost involved in setting up the proposed store, including the expenses associated with procuring the medicines.

Furthermore, the collected data will be leveraged to create an interactive dashboard. This dashboard will serve as a powerful tool for the consultancy firm's client, aiding them in making informed decisions regarding the selection of medicines for their store. The dashboard will facilitate a comparison of various parameters such as prices, customer ratings, and popularity, enabling the client to identify the best-suited medicines for their target market.

In conclusion, through the utilization of Beautifulsoup and data scraping techniques, this project aims to provide invaluable insights and recommendations to the consultancy firm's client interested in opening a homeopathic medicine store. The analysis of the scraped data will contribute to estimating the total cost of opening the store, while the interactive dashboard will empower the client to make informed decisions and optimize their selection of medicines for the store's success.